// server/validators/types.ts
export enum ValidatorMessages {
  Required = 'Field is required',
  Boolean = 'Field must be a boolean',
  Number = 'Field must be a number',
  String = 'Field must be a string',
  Array = 'Field must be an array',
  Email = 'Field must be an email',
  MongoId = 'Field must be a valid Mongo ID',
  Object = 'Field must be an object',
  NonEmptyString = 'Field must be a non-empty string',
}